requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
billiard==3.6.4.0
boto3==1.23.10
botocore==1.26.10
cached-property==1.5.2
celery==5.1.2
certifi==2022.5.18.1
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.13
django-storages==1.12.3
djangorestframework==3.13.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.3
incremental==21.3.0
isort==5.10.1
jmespath==0.10.0
kombu==5.1.0
msgpack==1.0.4
packaging==21.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.1
redis==4.3.3
requests==2.27.1
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.14.1
zipp==3.6.0
zope.interface==5.4.0
